backend/_order/serializers.py
from rest_framework import serializers

from backend._order import models


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Order
        fields = "__all__"
