requirements.txt
argcomplete==1.12.2
asgiref==3.3.1
attrs==20.3.0
boto3==1.17.33
botocore==1.20.33
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
coverage==5.5
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
durationpy==0.5
flake8==3.9.0
future==0.18.2
gunicorn==20.0.4
hjson==3.0.2
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
kappa==0.6.0
Markdown==3.3.4
mccabe==0.6.1
packaging==20.9
pep517==0.10.0
pip-tools==6.0.1
placebo==0.9.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.59.0
troposphere==2.6.4
urllib3==1.26.4
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
